backend-flask/requirements.txt
Flask==3.0.3
Flask-CORS==5.0.0
openai>=1.0.0
python-dotenv>=0.19.0